+++ LawnApp.h
@@ -2,6 +2,7 @@
 #define __LAWNAPP_H__
 
 #include "ConstEnums.h"
+#include "Sexy.TodLib/TodFoley.h"
 #include "SexyAppFramework/SexyApp.h"
 
 class Board;
@@ -95,7 +96,7 @@ public:
     int mAppRandSeed;                      //+0x870
     HICON mBigArrowCursor;                 //+0x874
     PopDRMComm *mDRM;                      //+0x878
-    int mSessionID;                        //+0x87C
+    intptr_t mSessionID;                   //+0x87C
     int mPlayTimeActiveSession;            //+0x880
     int mPlayTimeInactiveSession;          //+0x884
     BoardResult mBoardResult;              //+0x888
