+++ SexyAppFramework/FModLoader.cpp
@@ -24,7 +24,7 @@ FMOD_INSTANCE::FMOD_INSTANCE(const char *dllName) {
     if (!mModule) return;
 
 #define GETPROC(_x, _y)                                                                                                \
-    { CheckFModFunction(*((unsigned int *)&_x) = (unsigned int)GetProcAddress(mModule, _y), #_y); }
+    { CheckFModFunction(*((uintptr_t *)&_x) = (uintptr_t)GetProcAddress(mModule, _y), #_y); }
 
     GETPROC(FSOUND_SetBufferSize, "_FSOUND_SetBufferSize@4");
     GETPROC(FSOUND_SetHWND, "_FSOUND_SetHWND@4");
